Some improvements
When the acquired data are sent all together, the file saved by
yarp::dev::AudioPlayerWrapper is equal to the one saved py python wave
package. If the data are sent through the yarp buffered port as soon as
they arrive from the web page, the file saved by the yarp device is
"shrinked". To be investigated